+++ lib/consumer.js
@@ -2,8 +2,8 @@
  * @namespace Consumer
  */
 var amqp = require('amqplib');
-var logger = require('../helpers/logger')(true);
-var utils = require('../helpers/utils');
+var logger = require('./helpers/logger')(true);
+var utils = require('./helpers/utils');
 
 var config;
 var connected = false;
