+++ lib/helpers/logger.js
